+++ firmware/cm7/startup/boot/evkmimxrt1170_flexspi_nor_config.c
