+++ firmware/cm7/startup/boot/fsl_flexspi_nor_boot.c
