+++ firmware/cm7/startup/system_init/CMakeLists.txt
@@ -10,10 +10,6 @@ file(GLOB public_headers "*.h")
 target_sources(system_init
   PUBLIC
     ${public_headers}
-    evkmimxrt1170_flexspi_nor_config.c
-    evkmimxrt1170_flexspi_nor_config.h
-    fsl_flexspi_nor_boot.h
-    fsl_flexspi_nor_boot.c
   PRIVATE
     system_MIMXRT1176_cm7.c
     system_init_hook.cpp
