+++ test/cm7/hello_world_cm7.cpp
@@ -15,8 +15,8 @@
 #include <cstdio>
 
 extern "C" {
-#include "startup/system_init/evkmimxrt1170_flexspi_nor_config.h"
-#include "startup/system_init/fsl_flexspi_nor_boot.h"
+#include "startup/boot/evkmimxrt1170_flexspi_nor_config.h"
+#include "startup/boot/fsl_flexspi_nor_boot.h"
 extern const ivt image_vector_table;
 extern const BOOT_DATA_T g_boot_data;
 }
